+++ src/main/groovy/com/palantir/jacoco/JacocoCoverageExtension.groovy
@@ -16,8 +16,15 @@
 
 package com.palantir.jacoco
 
+import com.google.common.collect.ArrayListMultimap
+import com.google.common.collect.Multimap
+
 import java.util.regex.Pattern
 
+/**
+ * Provides configuration options for the {@link JacocoCoveragePlugin}, in particular the fileThreshold, classThreshold,
+ * packageThreshold, reportThreshold keywords for specifying coverage requirements.
+ */
 public class JacocoCoverageExtension {
 
     // For convenience in build scripts.
@@ -28,36 +35,72 @@ public class JacocoCoverageExtension {
     def LINE = CoverageType.LINE
     def METHOD = CoverageType.METHOD
 
-    public List<Closure<Double>> coverage = new ArrayList<>()
+    public Multimap<CoverageRealm, Closure<Double>> coverage = ArrayListMultimap.create();
+
+    public JacocoCoverageExtension() {
+        def mc = new ExpandoMetaClass(JacocoCoverageExtension, false, true)
+        mc.initialize()
+        this.metaClass = mc
+
+        // For each realm, add configuration specifiers (fileThreshold, classThreshold, etc) to the extension.
+        CoverageRealm.values().each { realm ->
+            mc."${realm.realmName}" = { double value ->
+                threshold(realm, value)
+            }
+
+            mc."${realm.realmName}" = { double value, CoverageType coverageType ->
+                threshold(realm, value, coverageType)
+            }
+
+            mc."${realm.realmName}" = { double value, String scope ->
+                threshold(realm, value, scope)
+            }
+
+            mc."${realm.realmName}" = { double value, CoverageType coverageType, String scope ->
+                threshold(realm, value, coverageType, scope)
+            }
+
+            mc."${realm.realmName}" = { double value, Pattern scopePattern ->
+                threshold(realm, value, scopePattern)
+            }
+
+            mc."${realm.realmName}" = { double value, CoverageType coverageType, Pattern scopePattern ->
+                threshold(realm, value, coverageType, scopePattern)
+            }
+        }
+    }
 
     /**
-     * Adds a minimum coverage threshold for all scopes and coverage types.
+     * Adds a minimum coverage threshold for all scopes and coverage types in the given realm.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      */
-    def threshold(double value) {
-        coverage.add({
+    def threshold(CoverageRealm realm, double value) {
+        coverage.put(realm, {
             CoverageType ct, String str -> value
         })
     }
 
     /**
-     * Adds a minimum coverage threshold for the given coverage type and for all scopes.
+     * Adds a minimum coverage threshold for the given coverage type and for all scopes in the given realm.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      * @param coverageType The type of JaCoCo coverage this threshold applies to.
      */
-    def threshold(double value, CoverageType coverageType) {
-        threshold(value, coverageType, ~/.*/)
+    def threshold(CoverageRealm realm, double value, CoverageType coverageType) {
+        threshold(realm, value, coverageType, ~/.*/)
     }
 
     /**
-     * Adds a minimum coverage threshold for all coverage types and the given scope name.
+     * Adds a minimum coverage threshold for all coverage types and the given scope name in the given realm.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      * @param coverageType The type of JaCoCo coverage this rule applies to.
      * @param scope The scope (e.g., file name, class name, package name, report name) that this threshold applies to.
      */
-    def threshold(double value, String scope) {
+    def threshold(CoverageRealm realm, double value, String scope) {
         CoverageType.values().each { coverageType ->
-            coverage.add({ CoverageType ct, String str ->
+            coverage.put(realm, { CoverageType ct, String str ->
                 if (coverageType == ct && str.equals(scope)) {
                     return value
                 }
@@ -67,13 +110,14 @@ public class JacocoCoverageExtension {
     }
 
     /**
-     * Adds a minimum coverage threshold for the given coverage type and scope (exact match).
+     * Adds a minimum coverage threshold for the given coverage type and scope name (exact match) in the given realm.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      * @param coverageType The type of JaCoCo coverage this rule applies to.
-     * @param scope The scope that this threshold applies to.
+     * @param scope The scope name that this threshold applies to.
      */
-    def threshold(double value, CoverageType coverageType, String scope) {
-        coverage.add({ CoverageType ct, String str ->
+    def threshold(CoverageRealm realm, double value, CoverageType coverageType, String scope) {
+        coverage.put(realm, { CoverageType ct, String str ->
             if (coverageType == ct && str.equals(scope)) {
                 return value
             }
@@ -82,14 +126,16 @@ public class JacocoCoverageExtension {
     }
 
     /**
-     * Adds a minimum coverage threshold for all coverage types and files scopes matching the given pattern.
+     * Adds a minimum coverage threshold for all coverage types and scopes (in the given realm) matching the given
+     * pattern.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      * @param coverageType The type of JaCoCo coverage this rule applies to.
      * @param scopePattern A regular expression specifying the names of the scopes that this threshold applies to.
      */
-    def threshold(double value, Pattern scopePattern) {
+    def threshold(CoverageRealm realm, double value, Pattern scopePattern) {
         CoverageType.values().each { coverageType ->
-            coverage.add({ CoverageType ct, String str ->
+            coverage.put(realm, { CoverageType ct, String str ->
                 if (coverageType == ct && str ==~ scopePattern) {
                     return value
                 }
@@ -99,41 +145,16 @@ public class JacocoCoverageExtension {
     }
 
     /**
-     * Adds a minimum coverage threshold for the given coverage type scopes matching the given pattern.
+     * Adds a minimum coverage threshold for the given coverage type and scopes matching the given pattern.
+     * @param realm The realm that this threshold applies to.
      * @param value The minimum required threshold, a number in [0,1].
      * @param coverageType The type of JaCoCo coverage this rule applies to.
      * @param scopePattern A regular expression specifying the names of the scopes that this threshold applies to.
      */
-    def threshold(double value, CoverageType coverageType, Pattern scopePattern) {
-        coverage.add({ CoverageType ct, String str ->
+    def threshold(CoverageRealm realm, double value, CoverageType coverageType, Pattern scopePattern) {
+        coverage.put(realm, { CoverageType ct, String str ->
             if (coverageType == ct && str ==~ scopePattern) {
                 return value
-            }
-            return 2.0 // Thresholds >1.0 are never considered, i.e. this rule will be ignored.
-        })
-    }
-
-    /**
-     * Exempts the given scope from coverage requirements.
-     * @param scope The name of the scope that is to be whitelisted.
-     */
-    def whitelist(String scope) {
-        coverage.add({ CoverageType ct, String str ->
-            if (str.equals(scope)) {
-                return 0.0 // Whitelist.
-            }
-            return 2.0 // Thresholds >1.0 are never considered, i.e. this rule will be ignored.
-        })
-    }
-
-    /**
-     * Exempts scopes matching the given pattern from coverage requirements.
-     * @param scopePattern A regular expression specifying the scope that is to be whitelisted.
-     */
-    def whitelist(Pattern scopePattern) {
-        coverage.add({ CoverageType ct, String str ->
-            if (str ==~ scopePattern) {
-                return 0.0 // Whitelist.
             }
             return 2.0 // Thresholds >1.0 are never considered, i.e. this rule will be ignored.
         })